refactor: isolate Tooltip stories
